requirements.txt
torch~=2.3.1+cu118
matplotlib~=3.10.1
numpy~=1.26.4
pandas~=2.2.3
pillow~=11.1.0